add customer class for layalty card

+++ shoppingBasket.java
@@ -28,8 +28,8 @@ public class ShoppingBasket {
     this.contents.clear();
   }
 
-  public int getTotalPrice() {
-    int total = 0;
+  public long getTotalPrice() {
+    long total = 0;
       for (Item item : this.contents) {
         total += item.getPrice();
       }
